fix translating test, add choice of translation provider
